add webui button kick
